uDynamicMenu.h - Dynamic menu generation from a comma-delimited list of commands  - DYNAMIC_MENU  - DYNAMIC_MENU_CATEGORIES  - DYNAMIC_MENU_COLOR_*  - DYNAMIC_MENU_VALUE_LOCX HELPER_MENU  - Made dynamic  - HELPER_MENU_CATEGORIES - comma separated list that defines the top level categories for the menu  - HELPER_MENU_VALUE_LOCX color configurability:  - HELPER_MENU_COLOR_LABEL  - HELPER_MENU_COLOR_VALUES  - HELPER_MENU_COLOR_ENABLED  - HELPER_MENU_COLOR_DISABLED Added _help documentation for helper commands Refactor / fixes

+++ src/engine/eChatBot.h
@@ -43,7 +43,6 @@ enum ResponseType
 
 struct eChatBotDataBase
 {
-
     struct MessagePart
     {
         MessagePart(tString message, REAL delay, bool useChatFlag = false, REAL chatFlagPercentage = 0)
@@ -81,7 +80,7 @@ struct eChatBotDataBase
         bool validateOutput = true;
 
         // SCHEDULED
-        int currentPartIndex = 0;
+        size_t currentPartIndex = 0;
         std::vector<MessagePart> messageParts;
 
         bool matchFound                           = false;
